+++ app/lib/shared/versions.dart
@@ -21,7 +21,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.08.16', // The current [runtimeVersion].
+  '2021.08.17', // The current [runtimeVersion].
   '2021.08.12',
   '2021.07.27',
 ];
@@ -55,7 +55,7 @@ bool shouldGCVersion(String version) =>
 final String runtimeSdkVersion = '2.13.4';
 final String toolStableDartSdkVersion = '2.13.4';
 final String toolStableFlutterSdkVersion = '2.2.3';
-final String toolPreviewDartSdkVersion = '2.14.0-377.4.beta';
+final String toolPreviewDartSdkVersion = '2.14.0-377.7.beta';
 final String toolPreviewFlutterSdkVersion = '2.5.0-5.1.pre';
 
 // Value comes from package:pana.