+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2021.08.16`.
+ * Bumped runtimeVersion to `2021.08.17`.
+ * Upgraded preview Dart analysis SDK to `2.14.0-377.7.beta`
+ * Upgraded pana to `0.20.0`.
  * NOTE: started to populate and use `PackageDocument.uploaderUserIds`
          TODO(deferred): remove `PackageDocument.uploaderEmails` in a future release
  * NOTE: added weekly (versioned) periodic task: `check-datastore-integrity`.